cleanup KDL IK
- fixed type of max_solver_iterations_
- directly read parameters into class members

+++ moveit_kinematics/kdl_kinematics_plugin/include/moveit/kdl_kinematics_plugin/kdl_kinematics_plugin.h
@@ -197,7 +197,7 @@ private:
   // Storage required for when the set of redundant joints is reset
   bool position_ik_;  // whether this solver is only being used for position ik
   const robot_model::JointModelGroup* joint_model_group_;
-  double max_solver_iterations_;
+  int max_solver_iterations_;
   double epsilon_;
   std::vector<JointMimic> mimic_joints_;
 };
